pyproject.toml
[tool.poetry]
name = "keys-cli"
version = "0.1.0"
description = "Prompt hotkey CLI"
authors = ["tyson <tyson.chan@proton.me>"]
license = "MIT"
readme = "README.md"
packages = [{include = "keys"}]

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = ">=6.0"
pynput = ">=1.7.6"
typer = {version = ">=0.9.0", extras = ["all"]}
pyperclip = {version = ">=1.8.2", optional = true}

[tool.poetry.scripts]
keys = "keys.__main__:cli"

[tool.poetry.extras]
copy = ["pyperclip"]

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "N",
    "UP",
    "B",
    "C4",
    "SIM",
    "RET",
]
ignore = [
    "SIM108",
    "SIM117",
    "E501",
    "B008",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
